fix wrong types in close job offer

+++ MyLearn/MyLearn/InputModels/CloseJobOffer.cs
@@ -3,8 +3,8 @@
     public class CloseJobOffer
     {
         public string JobOfferId { get; set; }
-        public float State { get; set; }
+        public int State { get; set; }
         public string StateDescription { get; set; }
-        public float Stars { get; set; }
+        public int Stars { get; set; }
     }
 }
